optimize JS
