+++ assets/main-end.js
@@ -70,6 +70,47 @@ function parallaxUpdate() {
   })
 }
 
-import cloudinary from 'cloudinary-core';
-let cl = cloudinary.Cloudinary.new( { cloud_name: "dikhrro6y"});
-cl.responsive();
+// Decode Email
+function decodeEmail(encodedString) {
+    // Holds the final output
+    var email = ""; 
+    
+    // Extract the first 2 letters
+    var keyInHex = encodedString.substr(0, 2);
+
+    // Convert the hex-encoded key into decimal
+    var key = parseInt(keyInHex, 16);
+
+    // Loop through the remaining encoded characters in steps of 2
+    for (var n = 2; n < encodedString.length; n += 2) {
+        
+        // Get the next pair of characters
+        var charInHex = encodedString.substr(n, 2)
+        
+        // Convert hex to decimal
+        var char = parseInt(charInHex, 16);
+
+        // XOR the character with the key to get the original character
+        var output = char ^ key;
+
+        // Append the decoded character to the output
+        email += String.fromCharCode(output);
+    }
+    return email;
+}
+
+function updateAnchor(el) {
+    var encoded = el.innerHTML;
+    var decoded = decodeEmail(encoded);
+    el.textContent = decoded;
+    el.href = 'mailto:' + decoded;
+}
+
+let emlElements = document.getElementsByClassName("eml-protected");
+emlElements.forEach((e) => {
+    updateAnchor(e)
+})
+
+// import cloudinary from 'cloudinary-core';
+// let cl = cloudinary.Cloudinary.new( { cloud_name: "dikhrro6y"});
+// cl.responsive();
