+++ data/startScreen.yml
@@ -1,2 +1,2 @@
-background: Group_4_fmbpw4.svg
+background: https://res.cloudinary.com/dikhrro6y/Group_4_fmbpw4.svg
 image: https://res.cloudinary.com/dikhrro6y/image/upload/v1664818982/dog.jpg
